refactor(FavoriteAcceptanceTest.java): Support를 추가하여 전역으로 사용할 메서드 분리
