+++ src/test/java/nextstep/subway/favorite/acceptance/FavoriteAcceptanceTest.java
@@ -1,12 +1,8 @@
-package nextstep.subway.favorite;
+package nextstep.subway.favorite.acceptance;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.AcceptanceTest;
 import nextstep.subway.auth.acceptance.AuthAcceptanceTest;
-import nextstep.subway.favorite.dto.FavoriteRequest;
-import nextstep.subway.favorite.dto.FavoriteResponse;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.station.StationAcceptanceTest;
@@ -14,20 +10,14 @@ import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-
-import java.util.List;
 
 import static nextstep.subway.auth.acceptance.AuthAcceptanceTest.토큰_발급_요청;
 import static nextstep.subway.line.acceptance.LineAcceptanceTest.지하철_노선_등록되어_있음;
 import static nextstep.subway.line.acceptance.LineSectionAcceptanceTest.지하철_노선에_지하철역_등록_요청;
 import static nextstep.subway.member.MemberAcceptanceTest.*;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DisplayName("즐겨찾기 관련 기능")
-public class FavoriteAcceptanceTest extends AcceptanceTest {
+class FavoriteAcceptanceTest extends FavoriteAcceptanceTestSupport {
 
     private LineResponse 신분당선;
     private LineResponse 이호선;
@@ -78,54 +68,5 @@ public class FavoriteAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> removeResponse = 즐겨찾기_삭제_요청(accessToken, getFavoriteId(createResponse));
         // then
         즐겨찾기_삭제됨(removeResponse);
-    }
-
-    public static long getFavoriteId(ExtractableResponse<Response> response) {
-        return Long.parseLong(response.header("Location").split("/")[2]);
-    }
-
-    public static ExtractableResponse<Response> 즐겨찾기_삭제_요청(String accessToken, Long favoriteId) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .pathParam("id", favoriteId)
-                .when().delete("/favorites/{id}")
-                .then().log().all().extract();
-    }
-
-    public static void 즐겨찾기_삭제됨(ExtractableResponse<Response> removeResponse) {
-        assertThat(removeResponse.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-
-    public static void 즐겨찾기_생성됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    public static ExtractableResponse<Response> 즐겨찾기_생성_요청(String accessToken, Long sourceId, Long targetId) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(new FavoriteRequest(sourceId, targetId))
-                .when().post("/favorites")
-                .then().log().all().extract();
-    }
-
-    public static ExtractableResponse<Response> 즐겨찾기_목록_조회_요청(String accessToken) {
-        return RestAssured
-                .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .auth().oauth2(accessToken)
-                .when().get("/favorites")
-                .then().log().all().extract();
-    }
-
-    public static void 즐겨찾기_조회됨(ExtractableResponse<Response> response, StationResponse source, StationResponse target) {
-        List<FavoriteResponse> favorites = response.jsonPath().getList(".", FavoriteResponse.class);
-        assertAll(
-                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
-                () -> assertThat(favorites.get(0).getSource().getName()).isEqualTo(source.getName()),
-                () -> assertThat(favorites.get(0).getTarget().getName()).isEqualTo(target.getName())
-        );
     }
 }
