+++ src/test/java/nextstep/subway/favorite/acceptance/FavoriteAcceptanceTestSupport.java
@@ -1,0 +1,67 @@
+package nextstep.subway.favorite.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.AcceptanceTest;
+import nextstep.subway.favorite.dto.FavoriteRequest;
+import nextstep.subway.favorite.dto.FavoriteResponse;
+import nextstep.subway.station.dto.StationResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+public class FavoriteAcceptanceTestSupport extends AcceptanceTest {
+    public static long getFavoriteId(ExtractableResponse<Response> response) {
+        return Long.parseLong(response.header("Location").split("/")[2]);
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_삭제_요청(String accessToken, Long favoriteId) {
+        return RestAssured
+                .given().log().all()
+                .auth().oauth2(accessToken)
+                .pathParam("id", favoriteId)
+                .when().delete("/favorites/{id}")
+                .then().log().all().extract();
+    }
+
+    public static void 즐겨찾기_삭제됨(ExtractableResponse<Response> removeResponse) {
+        assertThat(removeResponse.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    public static void 즐겨찾기_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_생성_요청(String accessToken, Long sourceId, Long targetId) {
+        return RestAssured
+                .given().log().all()
+                .auth().oauth2(accessToken)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(new FavoriteRequest(sourceId, targetId))
+                .when().post("/favorites")
+                .then().log().all().extract();
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_목록_조회_요청(String accessToken) {
+        return RestAssured
+                .given().log().all()
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .auth().oauth2(accessToken)
+                .when().get("/favorites")
+                .then().log().all().extract();
+    }
+
+    public static void 즐겨찾기_조회됨(ExtractableResponse<Response> response, StationResponse source, StationResponse target) {
+        List<FavoriteResponse> favorites = response.jsonPath().getList(".", FavoriteResponse.class);
+        assertAll(
+                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
+                () -> assertThat(favorites.get(0).getSource().getName()).isEqualTo(source.getName()),
+                () -> assertThat(favorites.get(0).getTarget().getName()).isEqualTo(target.getName())
+        );
+    }
+}
